test(mcp): put the flock helper behind a build tag
syscall.Flock does not exist on Windows, so vet failed to build the package and the Windows job has been red since yesterday evening.
